pyproject.toml
[tool.poetry]
name = "dankbot"
version = "2.1.0"
description = ""
authors = ["Ishan Manchanda <ishanmanchanda70@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "3.9.6"
imageio-ffmpeg = "^0.4.5"
imutils = "^0.5.4"
numba = "^0.53.1"
numpy = "^1.21.2"
opencv-python = "^4.5.3"
Pillow = "^8.3.2"
pyimgur = "^0.6.0"
python-dotenv = "^0.19.0"
python-telegram-bot = "^13.7"
pytz = "^2021.1"
requests = "^2.26.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
